+++ panel/project-files.rkt
@@ -4,8 +4,10 @@ NOTICE: modify from example in https://github.com/racket/gui/blob/master/gui-doc
 origin author: https://github.com/racket/gui/graphs/contributors
 modifier author: Lîm Tsú-thuàn(GitHub: @dannypsnl)
 |#
+(provide project-files%)
 
 (require mrlib/hierlist
+         file/glob
          "../project-manager.rkt")
 
 (define set-text-mixin
@@ -19,33 +21,41 @@ modifier author: Lîm Tsú-thuàn(GitHub: @dannypsnl)
       (send t erase)
       (send t insert str))))
 
-(provide project-files%)
+(define ignore-list
+  '(".git"
+    "compiled"
+    "doc"
+    ".DS_Store"))
+
 (define project-files%
   (class hierarchical-list% (init editor-panel)
     (define the-editor-panel editor-panel)
     ; new-item : create new item for a file or directory
     (define (new-item parent directory subpath)
       (let ([cur-path (build-path directory subpath)])
-        (if (file-exists? cur-path)
-            (let ([item (send parent new-item set-text-mixin)])
-              (send* item
-                [set-text (path->string subpath)]
-                [user-data (build-path directory subpath)]))
-            (let ([item (send parent new-list set-text-mixin)])
-              (send item set-text (path->string subpath))
-              (for ([i (directory-list cur-path)])
-                (new-item item cur-path i))))))
+        (when (not (glob-match? ignore-list subpath))
+          (match (file-or-directory-type cur-path #t)
+            ['file
+             (let ([item (send parent new-item set-text-mixin)])
+               (send* item
+                 [set-text (path->string subpath)]
+                 [user-data (build-path directory subpath)]))]
+            ['directory
+             (let ([item (send parent new-list set-text-mixin)])
+               (send item set-text (path->string subpath))
+               (for ([i (directory-list cur-path)])
+                 (new-item item cur-path i)))]))))
     ; Set the top level item, and populate it with an entry
     ; for each item in the directory.
     (define/public (set-directory dir)
-      (send this delete-item top-dir-list) ; remove previous top item
       (set! top-dir-list (send this new-list set-text-mixin))
       (send top-dir-list set-text (path->string dir))
       ; add new-item for each member of dir
-      (for ([i (directory-list dir)])
-        (new-item top-dir-list dir i))
+      (for ([sub (directory-list dir)])
+        (new-item top-dir-list dir sub))
       ;; open top dir-list by default
       (send top-dir-list open))
+
     (define/override (on-double-select i)
       (when (send i user-data) ;; when double-click a file, open it in editor
         (define path (send i user-data))
@@ -55,8 +65,19 @@ modifier author: Lîm Tsú-thuàn(GitHub: @dannypsnl)
               (send the-editor-panel open-in-new-tab path)))))
     ;;; init
     (super-new)
-    ; top item in hierlist
-    (define top-dir-list (send this new-list set-text-mixin))
+    (define top-dir-list #f)
     (send current-project listen
           (λ (new-dir)
             (send this set-directory new-dir)))))
+
+(module+ main
+  (define frame (new frame% [label "test: project files"]
+                     [width 300] [height 300]))
+  (define viewer
+    (new project-files% [parent frame]
+         [editor-panel #f]))
+
+  (send viewer set-directory (current-directory))
+
+  (send frame center)
+  (send frame show #t))